requirements.txt
pandas
scikit-learn
numpy
kafka-python
pyspark==3.4.1
snowflake-connector-python
boto3
pyyaml